+++ package.json
@@ -8,7 +8,6 @@
     "@material-ui/icons": "4.11.2",
     "@testing-library/jest-dom": "5.11.9",
     "@testing-library/react": "11.2.5",
-    "@types/react-router-dom": "5.1.7",
     "clsx": "1.1.1",
     "react": "16.14.0",
     "react-dom": "16.14.0",
@@ -87,6 +86,7 @@
     "@types/react": "16.14.0",
     "@types/react-dom": "16.9.11",
     "@types/react-redux": "7.1.16",
+    "@types/react-router-dom": "5.1.7",
     "@types/redux-logger": "3.0.8",
     "@types/testing-library__jest-dom": "5.9.5",
     "@typescript-eslint/eslint-plugin": "4.15.2",
