+++ src/App.test.tsx
@@ -2,7 +2,6 @@ import { render } from '@testing-library/react';
 import React from 'react';
 import { Provider } from 'react-redux';
 import store from 'States/root-store';
-
 import App from './App';
 
 test('renders App', () => {
