--- src/components/MainView/index.ts
@@ -1,3 +1,0 @@
-import MainView from './MainView';
-
-export default MainView;
