--- src/entities/laguage.ts
@@ -1,9 +1,0 @@
-interface Languages {
-  en: 'EN';
-  ru: 'RU';
-  de: 'DE';
-}
-
-type Language = keyof Languages;
-
-export default Language;
