+++ src/entities/travel-app.ts
@@ -1,0 +1,1 @@
+export type Language = 'en' | 'ru' | 'de';
