+++ src/index.scss
@@ -8,10 +8,6 @@ body {
 
 }
 
-#root {
-  height: 100vh;
-}
-
 code {
   font-family:
     source-code-pro,
