+++ src/redux-states/language-selector/actions.ts
@@ -1,4 +1,4 @@
-import Language from 'Entities/laguage';
+import { Language } from 'Entities/travel-app';
 import * as StateTypes from 'States/types';
 import * as t from './action-types';
