--- src/redux-states/language-selector/constants.ts
@@ -1,1 +1,0 @@
-export const NAME = 'languageSelector';
