+++ src/redux-states/language-selector/index.ts
@@ -1,9 +1,7 @@
 import * as actions from './actions';
-import * as constants from './constants';
 import reducer from './reducer';
 
 export default {
   actions,
-  constants,
   reducer,
 };
