+++ src/redux-states/language-selector/model.ts
@@ -1,4 +1,4 @@
-import Language from 'Entities/laguage';
+import { Language } from 'Entities/travel-app';
 
 export type IState = {
   language: Language;
