+++ src/redux-states/root-reducer.ts
@@ -2,7 +2,7 @@ import { combineReducers } from 'redux';
 import languageSelector from 'States/language-selector';
 
 const rootReducer = combineReducers({
-  [languageSelector.constants.NAME]: languageSelector.reducer,
+  languageSelector: languageSelector.reducer,
 });
 
 export default rootReducer;
